actualizacon publicaciones, sigue pendiente marketplace imagnes

+++ admin/publicaciones.php
@@ -294,6 +294,7 @@
                                           <th>Titulo</th>
                                           <th>Negocio</th>
                                           <th class="text-right">Fecha</th>
+                                          <th class="text-right">Estatus</th>
                                           <th class="text-right">Acción</th>
                                       </tr>
               </thead>
@@ -305,7 +306,10 @@ while($fila=mysqli_fetch_array($ejecutar)){
   $id       = $fila['ID_PUBLICACION'];
   $titu_pub  = $fila['PUB_TITULO'];
   $nom_neg   = $fila['NEG_NOMBRE'];
+  $estatus   = $fila['PUB_ESTATUS'];
+  $ingles   = $fila['NEG_INGLES'];
   $fecha_pub  = $fila['PUB_FECHA'];
+  $fecha_lim  = $fila['PUB_FECHA_LIMITE'];
 
 
 ?>
@@ -314,14 +318,35 @@ while($fila=mysqli_fetch_array($ejecutar)){
                                          <td><?php echo $titu_pub ?></td>
                                          <td><?php echo $nom_neg ?></td>
                                          <td><?php echo $fecha_pub ?></td>
+                                         <td><?php echo $estatus ?></td>
                                          <td width="14%">
-                                        <?php echo"
-                                        <a href='#' onclick='editar_pub($id), get_pub($id);' title='Modificar ' ><i class='btn-sm btn-success fa fa-refresh'></i></a>
-                                        <a href='publicacion_notificacion_fn.php?id=$id' onclick='pregunta()' name='tuformulario' title='Enviar notificacion esp ' ><i class='btn-sm btn-warning fa fa-bell'></i></a> 
-                                        <a href='publicacion_notificacion_fn_ing.php?id=$id' onclick='pregunta()' name='tuformulario' title='Enviar notificacion ing' ><i class='btn-sm btn-warning fa fa-bell'></i></a>  
-                                        <a href='estadisticas_xp.php?id=$id' title='Estadisticas ' ><i class='btn-sm btn-info fa fa-eye'></i></a>   
-                                        <a href='modificar_empresa_car_delete_fn.php?id=$id;' title='Eliminar ' ><i class='btn-sm btn-danger fa fa-trash'></i></a>   
-                                            "
+                                        <?php
+                                        if ($estatus == "A"){
+                                          if($ingles == "True"){
+                                            echo"
+                                            <a href='#' onclick='editar_pub($id), get_pub($id);' title='Modificar ' ><i class='btn-sm btn-success fa fa-refresh'></i></a>
+                                            <a href='publicacion_notificacion_fn.php?id=$id' onclick='pregunta()' name='tuformulario' title='Enviar notificacion esp ' ><i class='btn-sm btn-warning fa fa-bell'></i></a> 
+                                            <a href='publicacion_notificacion_fn_ing.php?id=$id' onclick='pregunta()' name='tuformulario' title='Enviar notificacion ing' ><i class='btn-sm btn-warning fa fa-bell'></i></a>  
+                                            <a href='estadisticas_xp.php?id=$id' title='Estadisticas ' ><i class='btn-sm btn-info fa fa-eye'></i></a>   
+                                            <a href='modificar_empresa_car_delete_fn.php?id=$id;' title='Eliminar ' ><i class='btn-sm btn-danger fa fa-trash'></i></a>   
+                                                ";
+
+                                          }if($ingles == "False"){
+                                            echo"
+                                          <a href='#' onclick='editar_pub($id), get_pub($id);' title='Modificar ' ><i class='btn-sm btn-success fa fa-refresh'></i></a>
+                                          <a href='publicacion_notificacion_fn.php?id=$id' onclick='pregunta()' name='tuformulario' title='Enviar notificacion esp ' ><i class='btn-sm btn-warning fa fa-bell'></i></a> 
+                                          <a href='estadisticas_xp.php?id=$id' title='Estadisticas ' ><i class='btn-sm btn-info fa fa-eye'></i></a>   
+                                          <a href='modificar_empresa_car_delete_fn.php?id=$id;' title='Eliminar ' ><i class='btn-sm btn-danger fa fa-trash'></i></a>   
+                                              ";
+                                          }
+                                        
+                                        }else{  echo "
+
+                                          <a href='#' onclick='editar_pub($id), get_pub($id);' title='Modificar ' ><i class='btn-sm btn-success fa fa-refresh'></i></a>
+                      
+                                          ";
+                                          }
+                                        
                                           ?>
 
                                       </tr>
